AI/requirements.txt
flask==3.0.0
torch
torchvision
facenet-pytorch
Pillow
numpy
